requirements.txt
facenet_pytorch==2.5.3
numpy==1.24.2
Pillow==10.3.0
torch==2.0.0
torchvision==0.15.1
